apps/litter-networker/src/python-utils/requirements.txt
boto3
botocore
Pillow
requests
geopy
tqdm
shapely

